# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tournament Countdown API",
    "version": "1.0.0",
    "description": "Simple countdown API with real-time updates."
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
  "paths": {
    "/api/settings": {
      "get": {
        "summary": "Get app settings",
        "responses": {
          "200": {
            "description": "Current settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "settings": { "$ref": "#/components/schemas/Settings" } },
                  "required": ["settings"]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update app settings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Settings" },
              "example": {
                "defaultDurationSeconds": 45,
                "presets": [30, 45, 60],
                "gpio": { "enabled": false, "relayPin": 17, "relayActiveHigh": true, "blinkHz": 2, "blinkDurationMs": 10000 }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "ok": { "type": "boolean" }, "settings": { "$ref": "#/components/schemas/Settings" } },
                  "required": ["ok", "settings"]
                }
              }
            }
          },
          "400": { "description": "Invalid settings" }
        }
      }
    },
    "/api/state": {
      "get": {
        "summary": "Get current countdown state",
        "responses": {
          "200": {
            "description": "State",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/State" }
              }
            }
          }
        }
      }
    },
    "/api/start": {
      "post": {
        "summary": "Start countdown",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "durationMs": { "type": "integer", "minimum": 0, "description": "Duration in milliseconds" }
                }
              },
              "example": { "durationMs": 45000 }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "state": { "$ref": "#/components/schemas/State" }
                  },
                  "required": ["ok", "state"]
                }
              }
            }
          }
        }
      }
    },
    "/api/reset": {
      "post": {
        "summary": "Reset countdown",
        "responses": {
          "200": {
            "description": "Reset",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "state": { "$ref": "#/components/schemas/State" }
                  },
                  "required": ["ok", "state"]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Settings": {
        "type": "object",
        "properties": {
          "defaultDurationSeconds": { "type": "integer", "minimum": 0 },
          "presets": { "type": "array", "items": { "type": "integer", "minimum": 0 } },
          "gpio": {
            "type": "object",
            "properties": {
              "enabled": { "type": "boolean" },
              "relayPin": { "type": "integer", "minimum": 0, "description": "BCM pin number" },
              "relayActiveHigh": { "type": "boolean" },
              "blinkHz": { "type": "number", "minimum": 0 },
              "blinkDurationMs": { "type": "integer", "minimum": 0 }
            },
            "required": ["enabled", "relayPin", "relayActiveHigh", "blinkHz", "blinkDurationMs"]
          }
        },
        "required": ["defaultDurationSeconds", "presets", "gpio"]
      },
      "State": {
        "type": "object",
        "properties": {
          "durationMs": { "type": "integer", "minimum": 0 },
          "endTime": { "type": "integer", "nullable": true, "description": "UTC ms timestamp when countdown ends" },
          "remainingMs": { "type": "integer", "minimum": 0 },
          "running": { "type": "boolean" }
        },
        "required": ["durationMs", "remainingMs", "running"]
      }
    }
  }
}